Moving RumConstants to common package
