+++ demo-app/src/main/java/io/opentelemetry/android/demo/OtelDemoApplication.kt
@@ -10,6 +10,7 @@ import android.app.Application
 import android.util.Log
 import io.opentelemetry.android.OpenTelemetryRum
 import io.opentelemetry.android.OpenTelemetryRumBuilder
+import io.opentelemetry.android.agent.setSlowRenderingDetectionPollInterval
 import io.opentelemetry.android.config.OtelRumConfig
 import io.opentelemetry.android.features.diskbuffering.DiskBufferingConfiguration
 import io.opentelemetry.api.common.AttributeKey.stringKey
@@ -19,6 +20,7 @@ import io.opentelemetry.api.trace.Tracer
 import io.opentelemetry.exporter.otlp.http.logs.OtlpHttpLogRecordExporter
 import io.opentelemetry.exporter.otlp.http.trace.OtlpHttpSpanExporter
 import io.opentelemetry.sdk.logs.internal.SdkEventLoggerProvider
+import java.time.Duration
 import kotlin.math.log
 
 const val TAG = "otel.demo"
