+++ instrumentation/fragment/src/main/java/io/opentelemetry/android/instrumentation/fragment/FragmentTracer.java
@@ -6,7 +6,7 @@
 package io.opentelemetry.android.instrumentation.fragment;
 
 import androidx.fragment.app.Fragment;
-import io.opentelemetry.android.config.RumConstants;
+import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.android.instrumentation.common.ActiveSpan;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.Span;
